Make withTags logging function public
Many teams within Block/Cash copy the wisp `withTags` function and use it to
wrap blocks of their own code in tags.

We have been working to make the existing tools and patterns for logging more consistent,
and making `withTags` public is a very minor change but one small step in that direction.